requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastadmin==0.1.41
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.5
isort==5.12.0
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pydantic==1.10.13
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
